# openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "Occupy API",
		"description": "\nAPI for a social + bookings platform for small businesses.\n\n- **Auth**: JWT Bearer (`Authorization: Bearer <token>`)\n- **Pagination**: Cursor-based via `cursor` and `limit` query params; responses include `items[]` and `nextCursor`.\n- **Errors**: Standardized envelope `{ \"error\": { \"code\", \"message\", \"details\" } }`.\n\nSee **/redoc** for a compact reference.\n",
		"version": "0.1.0"
	},
	"paths": {
...
		"/api/v1/posts/cliques/{cliqueId}/posts": {
			"post": {
				"tags": ["Posts"],
				"summary": "Create post in clique",
				"description": "Clique owners create posts to share updates with members.",
				"operationId": "create_post_route_api_v1_posts_cliques__cliqueId__posts_post",
				"security": [{ "HTTPBearer": [] }, { "BearerAuth": [] }],
				"parameters": [
					{
						"name": "cliqueId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid",
							"title": "Cliqueid"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PostCreate",
								"examples": {
									"announcement": {
										"summary": "Publish announcement",
										"value": {
											"content": "✨ Spring product launch this Friday at 5pm!",
											"status": "posted"
										}
									}
								}
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Post created",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/Post" },
								"example": {
									"id": "7415722e-4e4f-4f8b-8c44-2924f905a712",
									"cliqueId": "257c6140-3ab2-4e74-bac6-41b4ed9f8f2e",
									"authorUserId": "93d52d58-eac4-4e74-a69d-6410a1de0970",
									"content": "✨ Spring product launch this Friday at 5pm!",
									"status": "posted",
									"likesCount": 0,
									"commentsCount": 0,
									"likedByMe": false,
									"createdAt": "2024-04-01T12:00:00Z",
									"updatedAt": "2024-04-01T12:00:00Z"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"422": {
						"description": "Unprocessable Content",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					}
				}
			},
			"get": {
				"tags": ["Posts"],
				"summary": "List clique posts",
				"description": "Paginated posts visible to the current user, including engagement metadata.",
				"operationId": "list_clique_posts_route_api_v1_posts_cliques__cliqueId__posts_get",
				"security": [{ "HTTPBearer": [] }, { "BearerAuth": [] }],
				"parameters": [
					{
						"name": "cliqueId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid",
							"title": "Cliqueid"
						}
					},
					{ "$ref": "#/components/parameters/CursorParam" },
					{ "$ref": "#/components/parameters/LimitParam" }
				],
				"responses": {
					"200": {
						"description": "Posts page",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/CursorPagePosts" }
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			}
		},
		"/api/v1/posts/user/{userId}/posts": {
			"get": {
				"tags": ["Posts"],
				"summary": "List user posts",
				"description": "Paginated posts by a user visible to the current user, respecting privacy settings.",
				"operationId": "list_user_posts_route_api_v1_posts_user__userId__posts_get",
				"security": [{ "HTTPBearer": [] }, { "BearerAuth": [] }],
				"parameters": [
					{
						"name": "userId",
						"in": "path",
						"required": true,
						"schema": { "type": "string", "format": "uuid", "title": "Userid" }
					},
					{ "$ref": "#/components/parameters/CursorParam" },
					{ "$ref": "#/components/parameters/LimitParam" }
				],
				"responses": {
					"200": {
						"description": "User posts page",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/CursorPagePosts" }
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			}
		},
		"/api/v1/posts/{postId}": {
			"get": {
				"tags": ["Posts"],
				"summary": "Get post",
				"description": "Retrieve a post with reactions and visibility checks applied.",
				"operationId": "get_post_route_api_v1_posts__postId__get",
				"security": [{ "HTTPBearer": [] }, { "BearerAuth": [] }],
				"parameters": [
					{
						"name": "postId",
						"in": "path",
						"required": true,
						"schema": { "type": "string", "format": "uuid", "title": "Postid" }
					}
				],
				"responses": {
					"200": {
						"description": "Post detail",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/Post" }
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			},
			"patch": {
				"tags": ["Posts"],
				"summary": "Update post",
				"description": "Post authors or clique owners can edit content or status.",
				"operationId": "update_post_route_api_v1_posts__postId__patch",
				"security": [{ "HTTPBearer": [] }, { "BearerAuth": [] }],
				"parameters": [
					{
						"name": "postId",
						"in": "path",
						"required": true,
						"schema": { "type": "string", "format": "uuid", "title": "Postid" }
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PostUpdate",
								"examples": {
									"publish": {
										"summary": "Publish draft",
										"value": { "status": "posted" }
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Post updated",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/Post" }
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"422": {
						"description": "Unprocessable Content",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					}
				}
			},
			"delete": {
				"tags": ["Posts"],
				"summary": "Delete post",
				"description": "Delete a post as the author or clique owner.",
				"operationId": "delete_post_route_api_v1_posts__postId__delete",
				"security": [{ "HTTPBearer": [] }, { "BearerAuth": [] }],
				"parameters": [
					{
						"name": "postId",
						"in": "path",
						"required": true,
						"schema": { "type": "string", "format": "uuid", "title": "Postid" }
					}
				],
				"responses": {
					"204": { "description": "Post deleted" },
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			}
		},
		"/api/v1/posts/{postId}/like": {
			"post": {
				"tags": ["Posts"],
				"summary": "Like post",
				"description": "Toggle the current user's like on the specified post.",
				"operationId": "like_route_api_v1_posts__postId__like_post",
				"security": [{ "HTTPBearer": [] }, { "BearerAuth": [] }],
				"parameters": [
					{
						"name": "postId",
						"in": "path",
						"required": true,
						"schema": { "type": "string", "format": "uuid", "title": "Postid" }
					}
				],
				"responses": {
					"200": {
						"description": "Post with updated reactions",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/Post" },
								"example": {
									"id": "7415722e-4e4f-4f8b-8c44-2924f905a712",
									"likesCount": 12,
									"commentsCount": 3,
									"likedByMe": true
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			},
			"delete": {
				"tags": ["Posts"],
				"summary": "Unlike post",
				"description": "Remove the user's like from the post.",
				"operationId": "unlike_route_api_v1_posts__postId__like_delete",
				"security": [{ "HTTPBearer": [] }, { "BearerAuth": [] }],
				"parameters": [
					{
						"name": "postId",
						"in": "path",
						"required": true,
						"schema": { "type": "string", "format": "uuid", "title": "Postid" }
					}
				],
				"responses": {
					"200": {
						"description": "Post with updated reactions",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/Post" },
								"example": {
									"id": "7415722e-4e4f-4f8b-8c44-2924f905a712",
									"likesCount": 11,
									"commentsCount": 3,
									"likedByMe": false
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			}
		},
		"/api/v1/posts/{postId}/comments": {
			"post": {
				"tags": ["Posts"],
				"summary": "Create comment",
				"description": "Add a comment to a visible post.",
				"operationId": "create_comment_route_api_v1_posts__postId__comments_post",
				"security": [{ "HTTPBearer": [] }, { "BearerAuth": [] }],
				"parameters": [
					{
						"name": "postId",
						"in": "path",
						"required": true,
						"schema": { "type": "string", "format": "uuid", "title": "Postid" }
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CommentCreate",
								"examples": {
									"supportive": {
										"summary": "Friendly comment",
										"value": {
											"content": "Can't wait to see the new collection!"
										}
									}
								}
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Comment created",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/Comment" },
								"example": {
									"id": "b1d38d06-5801-4b69-90c6-74c4950c333a",
									"postId": "7415722e-4e4f-4f8b-8c44-2924f905a712",
									"userId": "93d52d58-eac4-4e74-a69d-6410a1de0970",
									"content": "Can't wait to see the new collection!",
									"createdAt": "2024-04-01T13:00:00Z"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"422": {
						"description": "Unprocessable Content",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					}
				}
			}
		},
		"/api/v1/posts/comments/{commentId}": {
			"delete": {
				"tags": ["Posts"],
				"summary": "Delete comment",
				"description": "Delete a comment authored by the user or managed by the clique owner.",
				"operationId": "delete_comment_route_api_v1_posts_comments__commentId__delete",
				"security": [{ "HTTPBearer": [] }, { "BearerAuth": [] }],
				"parameters": [
					{
						"name": "commentId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid",
							"title": "Commentid"
						}
					}
				],
				"responses": {
					"204": { "description": "Comment deleted" },
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"ContentFormat": {
				"type": "string",
				"enum": ["markdown"],
				"title": "ContentFormat"
			},
...
			"CursorPagePosts": {
				"properties": {
					"nextCursor": {
						"anyOf": [{ "type": "string" }, { "type": "null" }],
						"title": "Nextcursor"
					},
					"meta": {
						"anyOf": [
							{ "additionalProperties": true, "type": "object" },
							{ "type": "null" }
						],
						"title": "Meta"
					},
					"items": {
						"items": { "$ref": "#/components/schemas/Post" },
						"type": "array",
						"title": "Items"
					}
				},
				"type": "object",
				"required": ["items"],
				"title": "CursorPagePosts"
			},
...
			"HTTPValidationError": {
				"properties": {
					"detail": {
						"items": { "$ref": "#/components/schemas/ValidationError" },
						"type": "array",
						"title": "Detail"
					}
				},
				"type": "object",
				"title": "HTTPValidationError"
			},
...
			"Post": {
				"properties": {
					"contentFormat": {
						"$ref": "#/components/schemas/ContentFormat",
						"default": "markdown"
					},
					"content": { "type": "string", "title": "Content" },
					"status": {
						"$ref": "#/components/schemas/PostStatus",
						"default": "draft"
					},
					"id": { "type": "string", "format": "uuid", "title": "Id" },
					"cliqueId": {
						"type": "string",
						"format": "uuid",
						"title": "Cliqueid"
					},
					"authorUserId": {
						"type": "string",
						"format": "uuid",
						"title": "Authoruserid"
					},
					"deletedAt": {
						"anyOf": [
							{ "type": "string", "format": "date-time" },
							{ "type": "null" }
						],
						"title": "Deletedat"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"title": "Createdat"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"title": "Updatedat"
					},
					"likesCount": {
						"type": "integer",
						"title": "Likescount",
						"default": 0
					},
					"commentsCount": {
						"type": "integer",
						"title": "Commentscount",
						"default": 0
					},
					"likedByMe": {
						"type": "boolean",
						"title": "Likedbyme",
						"default": false
					}
				},
				"type": "object",
				"required": [
					"content",
					"id",
					"cliqueId",
					"authorUserId",
					"createdAt",
					"updatedAt"
				],
				"title": "Post"
			},
			"PostCreate": {
				"properties": {
					"contentFormat": {
						"$ref": "#/components/schemas/ContentFormat",
						"default": "markdown"
					},
					"content": { "type": "string", "title": "Content" },
					"status": {
						"$ref": "#/components/schemas/PostStatus",
						"default": "draft"
					}
				},
				"type": "object",
				"required": ["content"],
				"title": "PostCreate"
			},
...
			"PostUpdate": {
				"properties": {
					"content": {
						"anyOf": [{ "type": "string" }, { "type": "null" }],
						"title": "Content"
					},
					"status": {
						"anyOf": [
							{ "$ref": "#/components/schemas/PostStatus" },
							{ "type": "null" }
						]
					}
				},
				"type": "object",
				"title": "PostUpdate"
			},
...
			"ErrorEnvelope": {
				"type": "object",
				"required": ["error"],
				"properties": {
					"error": {
						"type": "object",
						"required": ["code", "message", "details"],
						"properties": {
							"code": { "type": "string", "example": "validation_error" },
							"message": {
								"type": "string",
								"example": "Invalid parent comment"
							},
							"details": {
								"type": "object",
								"additionalProperties": true,
								"example": {}
							}
						}
					}
				}
			}
		},
		"securitySchemes": {
			"HTTPBearer": { "type": "http", "scheme": "bearer" },
			"BearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		},
		"parameters": {
			"CursorParam": {
				"name": "cursor",
				"in": "query",
				"required": false,
				"schema": { "type": "string" },
				"description": "Opaque pagination cursor from previous response `nextCursor`."
			},
			"LimitParam": {
				"name": "limit",
				"in": "query",
				"required": false,
				"schema": {
					"type": "integer",
					"default": 20,
					"maximum": 100,
					"minimum": 1
				},
				"description": "Max items to return (default 20, max 100)."
			}
		},
...
	}
}